Makefile: Djinni target: depend on core lib sources without building it
Indirectly, through the core lib `sha512256` portability headers patch,
as it does not affect the API of the core lib, and Djinni should not
regenerate bindings when that patch is applied and has changed the
sources, so we'd rather apply it right away.

The Node bindings package, that depends on what Djinni generates, also
directly depends on the core lib to be built, but, since generating the
bindings does not, we simply don't build core lib just then yet.